Former-commit-id: a5a35d7b77051de86d280455e6063f1b79aa70af

+++ profiler/plugin/src/main/java/com/android/tools/profiler/ClassTransform.java
@@ -35,6 +35,7 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
+import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.zip.ZipEntry;
@@ -43,7 +44,7 @@ import java.util.zip.ZipOutputStream;
 
 /**
  * An abstract transform that provides a simple API for transforming all the possible classes.
- * The only abstract method is {@link #transform(InputStream, OutputStream)} which will be run
+ * The only abstract method is {@link #transform(InputStream, OutputStream, boolean)}which will be run
  * for every class. This class takes care of all the incremental and jar vs class details.
  */
 abstract class ClassTransform extends Transform {
@@ -51,8 +52,11 @@ abstract class ClassTransform extends Transform {
     @NonNull
     private final String mName;
 
-    public ClassTransform(String name) {
+    private final String[] mJarFilesToSkip;
+
+    public ClassTransform(String name, String[] jarFilesToSkip) {
         mName = name;
+        mJarFilesToSkip = jarFilesToSkip;
     }
 
     @NonNull
@@ -88,8 +92,35 @@ abstract class ClassTransform extends Transform {
         return String.format("%d_%s", number, name);
     }
 
+    //TODO Make this a regex, to cleanly support filtering.
+    private boolean isFileToSkip(File jarFile) {
+        for(String fileToSkip : mJarFilesToSkip) {
+            if (jarFile.getAbsolutePath().contains(fileToSkip)) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    /**
+     * Complete the previous class transform and building the inheritance hierarchy firstly
+     * (set buildHierarchyOnly true) then instrument on User Class (set buildHierarchyOnly false)
+     *
+     * The reason to do this is before instrumenting on user class, we need to create the inheritance
+     * hierarchy firstly
+     *
+     * @param invocation TransformInvocation
+     * @throws InterruptedException
+     * @throws IOException
+     */
     @Override
     public void transform(@NonNull TransformInvocation invocation)
+            throws InterruptedException, IOException {
+        transformHelper(invocation, true);
+        transformHelper(invocation, false);
+    }
+
+    public void transformHelper(@NonNull TransformInvocation invocation, boolean buildHierarchyOnly)
             throws InterruptedException, IOException {
         assert invocation.getOutputProvider() != null;
 
@@ -102,20 +133,26 @@ abstract class ClassTransform extends Transform {
                 File outputJar = invocation.getOutputProvider().getContentLocation(
                         name, jarInput.getContentTypes(), jarInput.getScopes(), Format.JAR);
 
+                // Check if this is a file we should skip or not.
+                if (isFileToSkip(inputJar)) {
+                    FileUtils.copyFile(inputJar,outputJar);
+                    continue;
+                }
+
                 if (invocation.isIncremental()) {
                     switch (jarInput.getStatus()) {
                         case NOTCHANGED:
                             break;
                         case ADDED:
                         case CHANGED:
-                            transformJar(inputJar, outputJar);
+                            transformJar(inputJar, outputJar, buildHierarchyOnly);
                             break;
                         case REMOVED:
                             FileUtils.delete(outputJar);
                             break;
                     }
                 } else {
-                    transformJar(inputJar, outputJar);
+                    transformJar(inputJar, outputJar, buildHierarchyOnly);
                 }
             }
             for (DirectoryInput di : ti.getDirectoryInputs()) {
@@ -134,7 +171,7 @@ abstract class ClassTransform extends Transform {
                             case CHANGED:
                                 for (File in : FileUtils.getAllFiles(inputFile)) {
                                     File out = toOutputFile(outputDir, inputDir, in);
-                                    transformFile(in, out);
+                                    transformFile(in, out, buildHierarchyOnly);
                                 }
                                 break;
                             case REMOVED:
@@ -146,14 +183,14 @@ abstract class ClassTransform extends Transform {
                 } else {
                     for (File in : FileUtils.getAllFiles(inputDir)) {
                         File out = toOutputFile(outputDir, inputDir, in);
-                        transformFile(in, out);
+                        transformFile(in, out, buildHierarchyOnly);
                     }
                 }
             }
         }
     }
 
-    private void transformJar(File inputJar, File outputJar) throws IOException {
+    private void transformJar(File inputJar, File outputJar, boolean buildHierarchyOnly) throws IOException {
         Files.createParentDirs(outputJar);
         try (FileInputStream fis = new FileInputStream(inputJar);
              ZipInputStream zis = new ZipInputStream(fis);
@@ -163,7 +200,7 @@ abstract class ClassTransform extends Transform {
             while (entry != null) {
                 zos.putNextEntry(new ZipEntry(entry.getName()));
                 if (!entry.isDirectory() && entry.getName().endsWith(".class")) {
-                    transform(zis, zos);
+                    transform(zis, zos, buildHierarchyOnly);
                 } else {
                     ByteStreams.copy(zis, zos);
                 }
@@ -172,11 +209,11 @@ abstract class ClassTransform extends Transform {
         }
     }
 
-    private void transformFile(File inputFile, File outputFile) throws IOException {
+    private void transformFile(File inputFile, File outputFile, boolean buildHierarchyOnly) throws IOException {
         Files.createParentDirs(outputFile);
         try (FileInputStream fis = new FileInputStream(inputFile);
              FileOutputStream fos = new FileOutputStream(outputFile)) {
-            transform(fis, fos);
+            transform(fis, fos, buildHierarchyOnly);
         }
     }
 
@@ -185,5 +222,5 @@ abstract class ClassTransform extends Transform {
         return new File(outputDir, FileUtils.relativePath(inputFile, inputDir));
     }
 
-    protected abstract void transform(InputStream in, OutputStream out) throws IOException;
+    protected abstract void transform(InputStream in, OutputStream out, boolean buildHierarchyOnly) throws IOException;
 }
